agent_custom/requirements.txt
# Core Framework
python-fasthtml>=0.5.0
flask>=3.0.0
flask-cors>=4.0.0

# AI & LLM
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-deepseek>=0.1.0
langgraph>=0.0.30
google-generativeai>=0.3.0

# Data & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP & Async
httpx>=0.26.0
requests>=2.31.0
aiofiles>=23.0.0

# Video Processing
yt-dlp>=2024.0.0

# Database
sqlite-utils>=3.35.0

# Utilities
python-multipart>=0.0.6
pandas>=2.0.0
pyyaml>=6.0.0

